fix: add tests
* test: add logger tests

* test: add tests for data source handling

* refactor: move tests to a separate directory

Jest was testing the fixtures, and moving things was faster than trying
to figure out why. :)

* fix: improve error handling

While writing tests, realized that there were uncaught Promise
rejections and that part of the error handling chain was apparently
unreachable. This fixes that.

* test: add Code Climate coverage reporting

* test: enable testing during CI

* refactor: make the gateway more testable

Express is hard to mock, so by splitting this into an app-specific and
server-specific set of files, we can bypass mocking Express altogether.

* refactor: update cleanup for easier testing

The cleanup process is now its own helper file. Made the cleanup
synchronous to simplify the program flow and make testing easier.
Removed some unused imports.

close #22

* refactor: map + filter -> filter